Added components folder to separate out index.jsx
